backend/requirements.txt
flask
flask-cors
requests
python-dotenv
supabase
sentence-transformers